#30 Исправлена логика метода удаления категории сервиса category-service

+++ core/category-service/src/main/java/malyshev/egor/service/admins/AdminCategoryServiceImpl.java
@@ -1,9 +1,11 @@
 package malyshev.egor.service.admins;
 
 import lombok.RequiredArgsConstructor;
+import malyshev.egor.InteractionApiManager;
 import malyshev.egor.dto.category.CategoryDto;
 import malyshev.egor.dto.category.NewCategoryDto;
 import malyshev.egor.dto.category.UpdateCategoryRequest;
+import malyshev.egor.exception.CategoryHasEventsException;
 import malyshev.egor.exception.NotFoundException;
 import malyshev.egor.model.Category;
 import malyshev.egor.repository.CategoryRepository;
@@ -17,6 +19,7 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional(readOnly = true)
 public class AdminCategoryServiceImpl implements AdminCategoryService {
     private final CategoryRepository repo;
+    private final InteractionApiManager interactionApiManager;
 
     // ADMIN
     @Override
@@ -47,6 +50,9 @@ public class AdminCategoryServiceImpl implements AdminCategoryService {
     public void delete(long id) {
         if (!repo.existsById(id)) {
             throw new NotFoundException("Category with id=" + id + " was not found");
+        }
+        if (interactionApiManager.existsEventsByCategoryId(id)) {
+            throw new CategoryHasEventsException("Category with id=" + id + " has linked events and cannot be deleted")
         }
         repo.deleteById(id);
     }
